[pruning][core][feature] Add model conversion for structured pruning
Summary:

This PR implements `convert` in the sparsifier. This function will return a model with pruned linear/conv2d layers.

Before, this functionality existed as a separate object in PrunedFuser.

The new convert function works by finding instances of specific patterns in the graph we get from
symbolically tracing the model.

Once we've found a match with a pattern, we then run that patterns custom conversion function to resize the weights appropriately. The conversion function also takes care of bias propogation.

We've implemented conversion functions for conv2d and linear patterns, in `convert_functions.py`.
There may be some gaps in the patterns compared to the previous implementation, but the conversion functionality should be the same.

Also migrated the tests from the previous PrunedFuser.

Test Plan:
```
python test/test_ao_sparsity.py -- TestBaseStructuredSparsifierConvert
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 32db89a5ab343d782f31167f94e4577adf49f2cc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/88625